SSA-183 Siempo shows information about notifications in tray.  - > SSA-200 Create UI for Notification like Default notification  - > SSA-207 Play store alpha release crash  - > SSA -136 Intention field is not getting cleared after selecting any application.  ->  SSA -137 Every time on clearing a notification the tray hide back and need to drag down to clear another notification,   -> SSA-214, 215 related to SSA-157 crash.

+++ MiniumApps/launcher3/src/main/java/co/siempo/phone/notification/remove_notification_strategy/SingleIteamDelete.java
@@ -9,16 +9,21 @@ import co.siempo.phone.notification.Notification;
  * Created by tkb on 2017-04-03.
  */
 
-@SuppressWarnings("ALL")
+
 public class SingleIteamDelete implements DeleteStrategy {
     @Override
     public void delete(Notification notification) {
-        TableNotificationSms notificationSms = DBUtility.getNotificationDao().queryBuilder()
-                .where(TableNotificationSmsDao.Properties._contact_title.eq(notification.getNumber()),
-                        TableNotificationSmsDao.Properties.Notification_type.eq(notification.getNotificationType()),
-                        TableNotificationSmsDao.Properties.Id.eq(notification.getId()))
-                .unique();
-
-        DBUtility.getNotificationDao().delete(notificationSms);
+        try {
+            TableNotificationSms notificationSms = DBUtility.getNotificationDao().queryBuilder()
+                    .where(TableNotificationSmsDao.Properties._contact_title.eq(notification.getNumber()),
+                            TableNotificationSmsDao.Properties.Notification_type.eq(notification.getNotificationType()),
+                            TableNotificationSmsDao.Properties.Id.eq(notification.getId()))
+                    .unique();
+            if (notificationSms != null) {
+                DBUtility.getNotificationDao().delete(notificationSms);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 }
